Fix: Convert type Integer into Double for Hour in SessionRecords

+++ backend/backend/src/main/java/com/tutor_management/backend/modules/finance/dto/response/InvoiceItem.java
@@ -10,7 +10,7 @@ public class InvoiceItem {
     private String date; // DD/MM/YYYY
     private String description; // "Buổi học tiếng Anh"
     private Integer sessions;
-    private Integer hours;
+    private Double hours;
     private Long pricePerHour;
     private Long amount;
 }
